requirements.txt
streamlit
google-generativeai
python-dotenv
groq

SpeechRecognition
opencv-python
fpdf2
